requirements.txt
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
PyJWT==2.10.0
pymongo==4.10.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
httpcore~=1.0.7
httpx~=0.27.2
anyio~=4.6.2.post1
pip~=24.3.1
distro~=1.9.0
pillow~=11.0.0
filelock~=3.16.1
typing_extensions~=4.12.2
python-dotenv~=1.0.1